+++ locales/en/index.yml
@@ -1598,11 +1598,13 @@ services:
       title: "How do you want to use services on IO?"
       body: "IO is simple because there's nothing to set up: if a service has something important to tell you, it can send you a message. Alternatively, you can activate services one by one."
       quickConfig:
+        value: "quick setup"
         title: "Use quick setup"
         body:
           text1: "Keep all services active, only those who have something important to tell you will contact you."
           text2: "You can always deactivate services that do not interest you."
       manualConfig:
+        value: "manual setup"
         title: "Configure manually"
         body:
           text1: "All services will be deactivated and you will have to enable them one by one to receive messages."
