+++ ts/components/services/ServicesSearch.tsx
@@ -154,7 +154,6 @@ class ServicesSearch extends React.PureComponent<Props, State> {
         isRefreshing={isFiltering}
         onRefresh={onRefresh}
         onSelect={this.handleOnServiceSelect}
-        isLongPressEnabled={false}
       />
     ) : (
       <SearchNoResultMessage errorType="NoResultsFound" />
